Minor refactoring of comments meta logic
